% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFishTaxa-package.R
\name{rFishTaxa-package}
\alias{rFishTaxa-package}
\alias{rFishTaxa}
\title{Eschmeyer's Catalog of Fishes Mining with R}
\description{
Eschmeyer's Catalog of Fishes \url{https://www.calacademy.org/scientists/projects/catalog-of-fishes} is the authoritative reference for taxonomic fish names, featuring a searchable online database.
This package helps users to obtain valid taxonomic fish information.
}
\references{
\itemize{
\item Fricke, R., Eschmeyer, W. N. & R. van der Laan (eds) 2022. ESCHMEYER'S CATALOG OF FISHES: GENERA, SPECIES, REFERENCES.. Electronic version accessed dd mmm 2022. \url{http://researcharchive.calacademy.org/research/ichthyology/catalog/fishcatmain.asp}
\item van der Laan, R. & Fricke, R. 2022. FAMILY-GROUP NAMES.Electronic version accessed dd mmm 2022. \url{http://www.calacademy.org/scientists/catalog-of-fishes-family-group-names}
\item van der Laan, R., Fricke, R. & Eschmeyer, W. N. (eds) 2022. ESCHMEYER'S CATALOG OF FISHES: CLASSIFICATION.Electronic version accessed dd mmm 2022. \url{http://www.calacademy.org/scientists/catalog-of-fishes-classification/}
\item Fricke, R. & Eschmeyer, W. N. 2022. JOURNALS.Electronic version accessed dd mmm 2022.\url{http://researcharchive.calacademy.org/research/ichthyology/catalog/journals.asp}
\item Fricke, R. & Eschmeyer, W. N. 2022. GUIDE TO FISH COLLECTIONS. Electronic version accessed dd mmm 2022.\url{http://researcharchive.calacademy.org/research/ichthyology/catalog/collections.asp}
\item Fricke, R., Eschmeyer, W. N. & Fong, J. D. 2022 SPECIES BY FAMILY/SUBFAMILY.Electronic version accessed dd mmm 2022.\url{http://researcharchive.calacademy.org/research/ichthyology/catalog/SpeciesByFamily.asp}
\item Fricke, R., Eschmeyer, W. N. & van der Laan, R. 2022. Eschmeyer's Catalog of Fishes online. \url{https://www.researchgate.net/project/Eschmeyers-Catalog-of-Fishes-online}
}
}
\keyword{package}
