% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_catalog.R
\name{search_cas}
\alias{search_cas}
\title{Searching Eschmeyer's Catalog of Fishes}
\format{
The following data frame of the fields to be returned are recognized:
\itemize{
\item query : family, genus, and species matched.
\item species(genus)_author : species(genus) and author(s) returned.
\item species(genus) : Species(genus) returned.
\item author : Author(s) returned.
\item family : Family(subfamily) of species.
\item status : Current status including validation, synonym, and uncertainty.
}
}
\usage{
search_cas(query, type)
}
\arguments{
\item{query}{\code{string} One and more queries, support family, genus, and species.}

\item{type}{\code{string} There is one required parameter, type = c("genus_family","species_family","species_genus","species").For example, `genus_family` for genera in a family,`species_family` for species in a family, `species_genus` for species in a genus, and `species` for species related to a species.}
}
\value{
Data frame.
}
\description{
This function get fields for fetching valid species information from the `Eschmeyer's Catalog of Fishes`.
}
\details{
See for website \url{http://researcharchive.calacademy.org/research/ichthyology/catalog/fishcatmain.asp} details.
}
\examples{
\dontrun{
# For genera in a family:
r1 <- search_cas(query = c("cyprinidae","balitoridae"),type = "genus_family")

# For species in a family:
r2 <- search_cas(query = "balitoridae",type = "species_family")

# For species in a genus:
r3 <- search_cas(query = "Brachyplatystoma",type = "species_genus")

# For species related to a species or subspecies:
r4 <- search_cas(query = c("Hoplias malabaricus", "Brachyplatystoma filamentosum"),type = "species")

}
}
\references{
Fricke, R., Eschmeyer, W. N. & R. van der Laan (eds) 2022. ESCHMEYER'S CATALOG OF FISHES: GENERA, SPECIES, REFERENCES. Electronic version accessed dd mmm 2022. \url{http://researcharchive.calacademy.org/research/ichthyology/catalog/fishcatmain.asp}
}
\author{
Liuyong Ding \email{ly_ding@126.com}
}
