% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_family.R
\name{species_family}
\alias{species_family}
\title{Obtaining species of fishes by family/subfamily periodically}
\format{
The following data frame of the fields to be returned are recognized:
\itemize{
\item class : Class of fishes returned.
\item order : Order of fishes returned.
\item family : Family of fishes returned.
\item subfamily : Subfamily of fishes returned.
\item available_genera : The number of available genera.
\item valid_genera : The number of valid genera.
\item available_species : The number of available species.
\item valid_species : The number of valid species.
\item ...
}
}
\usage{
species_family()
}
\value{
Data frame.
}
\description{
This function get genera/species of fishes by family/subfamily in real time.
}
\details{
See for website \url{https://researcharchive.calacademy.org/research/ichthyology/catalog/SpeciesByFamily.asp} details.
}
\examples{
\dontrun{
# Genera/Species of Fishes by Family/Subfamily through 2022:

db <- species_family()

# Each matching row by family

library(magrittr)
r <- search_cas(query = c("Anguilla nebulosa", "Clupisoma sinense"),type = "species")
r \%>\% dplyr::left_join(db[,1:4],by = "family")

}
}
\references{
\url{https://github.com/curso-r/lives/blob/6e042cf5f0bae0d957ff487a39b6601ac6f514ea/drafts/20210915_webscraping_ichthyology.R}
}
\author{
Liuyong Ding \email{ly_ding@126.com}
}
